server/requirements.txt
wget
argparse
tqdm
redis
flask
flask-cors
praw
python-dotenv